docs(falcon-ui): add custom props table component
