+++ packages/falcon-ui/docs/index.ts
@@ -1,4 +1,5 @@
-export { default as PropsTable } from './PropsTable';
+export { default as CustomPropsTable } from './CustomPropsTable';
 export { default as ResponsivePropsTable } from './ResponsivePropsTable';
 
 export { default as ThemePreview } from './ThemePreview';
+export * from './themeOverrides';
